+++ src/trustfake/models/torch/__init__.py
@@ -1,3 +1,13 @@
+from .clip import CLIPZeroShotClassifier, build_text_prototypes, clip_zeroshot
 from .resnet import resnet18, resnet34, resnet50, resnet101, resnet152
 
-__all__ = ["resnet18", "resnet34", "resnet50", "resnet101", "resnet152"]
+__all__ = [
+    "resnet18",
+    "resnet34",
+    "resnet50",
+    "resnet101",
+    "resnet152",
+    "CLIPZeroShotClassifier",
+    "build_text_prototypes",
+    "clip_zeroshot",
+]